le type finis avec un petit pb a corriger
une erreur de UnhandledPromiseRejectionWarning: dans le post.  le post est tout de meme fait dans la bdd

+++ controllers/TypeController.ts
@@ -21,40 +21,51 @@ export class TypeController{
     }
 
     public async getAll(limit:number,offset:number):Promise<TypeInstance[] | null>{
-        return this.Type.findAll({
+        return await this.Type.findAll({
             limit,
             offset
         });
     }
 
-    public async getById(id:number):Promise<TypeInstance|null>{
-        return this.Type.findOne({where: {
+    public async getById(id:string):Promise<TypeInstance|null>{
+        return await this.Type.findOne({where: {
                 id
             }});
     }
 
     public async removeById(id:string):Promise<Boolean>{
-        const type = this.getById(parseInt(id));
-        //if(type===undefined) return 0;
-        const destroy=this.Type.destroy({
-            where:{
-                id
+        const typeToDelete = await this.getById(id);
+        if(typeToDelete === null)
+        {
+            return false;
+        }
+        else
+        {
+            try
+            {
+                await this.Type.destroy({
+                    where:{
+                        id: typeToDelete.id
+                    }
+                });
+                return true;
             }
-        });
-        //console.log(destroy);
-        //if()
-
-        return false;
+            catch (err)
+            {
+                console.error(err);
+                return false;
+            }
+        }
     }
 
     public async addType(props: TypeCreationProps): Promise<TypeInstance | null> {
-        return this.Type.create({
+        return await this.Type.create({
             ...props
         });
     }
 
-    public async update(props:TypeInstance):Promise<TypeInstance | null>{
-        const typeUpdate = await this.getById(props.id);
+    public async update(options: { name: any; id: any }):Promise<TypeInstance | null>{
+        const typeUpdate = await this.getById(options.id.toString());
 
         if(typeUpdate === null)
         {
@@ -64,14 +75,13 @@ export class TypeController{
         {
             return await typeUpdate.update({
 
-                name: props.name
+                name: options.name
             }, {
                 where: {
-                    id: props.id
+                    id: options.id
                 }
             });
         }
-        return null;
     }
 
 }
